# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ML API",
    "description": "\n    Machine Learning API\n\n    - Upload datasets\n    - Register and train ML models\n    - Evaluate and rank models\n    - Make predictions\n    - List models and view reports\n    ",
    "contact": {
      "name": "AM",
      "url": "https://github.com/a1mart/ml",
      "email": "aidan3martin@gmail.com"
    },
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" },
    "version": "1.0.0"
  },
  "paths": {
    "/datasets/upload/": {
      "post": {
        "tags": ["Datasets"],
        "summary": "Upload Dataset",
        "description": "Upload a dataset file",
        "operationId": "upload_dataset_datasets_upload__post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_datasets_upload__post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Upload Dataset Datasets Upload  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/datasets/{dataset_name}": {
      "get": {
        "tags": ["Datasets"],
        "summary": "Get Dataset Info",
        "description": "Get detailed information about a dataset",
        "operationId": "get_dataset_info_datasets__dataset_name__get",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Dataset Name" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DatasetInfo" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Datasets"],
        "summary": "Delete Dataset",
        "description": "Delete a dataset",
        "operationId": "delete_dataset_datasets__dataset_name__delete",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Dataset Name" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/testing/complete-workflow": {
      "post": {
        "tags": ["Testing Workflow"],
        "summary": "Test Complete Workflow",
        "description": "Test the complete ML workflow from dataset creation to inference",
        "operationId": "test_complete_workflow_testing_complete_workflow_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WorkflowTest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/testing/cleanup/{dataset_name}": {
      "get": {
        "tags": ["Testing Workflow"],
        "summary": "Cleanup Test Data",
        "description": "Clean up test data and models",
        "operationId": "cleanup_test_data_testing_cleanup__dataset_name__get",
        "parameters": [
          {
            "name": "dataset_name",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Dataset Name" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_dataset_datasets_upload__post": {
        "properties": {
          "file": { "type": "string", "format": "binary", "title": "File" }
        },
        "type": "object",
        "required": ["file"],
        "title": "Body_upload_dataset_datasets_upload__post"
      },
      "DatasetInfo": {
        "properties": {
          "filename": { "type": "string", "title": "Filename" },
          "columns": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Columns"
          },
          "rows": { "type": "integer", "title": "Rows" },
          "size_bytes": { "type": "integer", "title": "Size Bytes" },
          "file_type": { "type": "string", "title": "File Type" },
          "target_suggestions": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Target Suggestions"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "columns",
          "rows",
          "size_bytes",
          "file_type",
          "target_suggestions"
        ],
        "title": "DatasetInfo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "WorkflowTest": {
        "properties": {
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name",
            "default": "test_dataset"
          },
          "n_samples": {
            "type": "integer",
            "title": "N Samples",
            "default": 1000
          },
          "n_features": {
            "type": "integer",
            "title": "N Features",
            "default": 20
          },
          "n_classes": {
            "type": "integer",
            "title": "N Classes",
            "default": 2
          },
          "models_to_test": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Models To Test",
            "default": ["LogisticRegression", "RandomForest", "XGBoost"]
          }
        },
        "type": "object",
        "title": "WorkflowTest"
      }
    }
  }
}